move video data ui to video module
